doh.lua: revamp for conformance with final draft
DNS responses are scanned for TTLs in order to set the HTTP response
lifetime.

The DNS server address is the same as the HTTP server address -
this is so that doh101 can run on a DNS server that uses views.

The content-type checking and other error checking should
be better now.